Added regular expressions (compiled) to a new non-generic abstract base class `EventMapper` for the `AbstractEventMapper<T>` in order to use those regexes inside the `AbstractEventMapper<T>` and the `DefaultEventMapper<T>`.
The implementation leads to an `ArgumentException` when using a `WillSerialize(IEventSerializer<T>)` with incorrect `Type` property format in the serializer or when using `WillSerialize(string)` with incorrect format.

I found no good way to implement just a "warning" without having to introduce an `ILogger<T>` and a default constructor to the `AbstractEventMapper<T>` class (which I don't want at the moment).