Refactored a lot, so that in the end only an absolute minimum of the code must/can be source-generated.
On top, this code does not have to be source-generated but can be implemented manually (because it's simple) or with some reflection (maybe as easy-to-use case - but there should be some analyzer message or warning, that the source-generated part is better and simpler).

The source generation is now driving the mapping (it creates the mapper implementation etc.).