More Projection... I need to think about how simple I can have a default projectionn. And then I want to have an option to do custom projections on existing event store data. (Not only for one stream id, but for all... maybe)...
How simple can I make it, to have multiple Projections for an Aggregate? e.g. one for the default Ralational-Database-State, then one for the RabbitMQ publishing (hopefully a generic one) and then maybe another to reflect the state into another model.

Is ist better to create the projections from the beginning, to allow inheritance etc. or is it better to only generate an IProjector with relevant calls that throws exception at the beginning when any projection implementation is missing... then allow the source-generator to replace the exceptions with the correct calls, as soon as there are AbstractProjections implemented for the relevant Events...?
Last idea seams good.

Also I could generate a default Projection that will just serialize the aggregates state into a json blob in the EventStoreDbContext as (table) "States". The aggregate could be found by it's stream id. It would be like a snapshot and pretty much everything that might be needed for a simple application.