Implemented `IStateStore`/`StateStore` as addition to the `IEventStore`/`EventStore`, which should simply store the current states of the aggregates.
The `IStateStore` will later be available to the developer through a - to be defined - `IReadRepository<T>`, so it could be used in production when multiple fields of an aggregate need to be checked for uniqueness etc.

This store is just a simple starting point for the developer and *should not*/*must not* replace the custom "lookup repositories" and the corresponding projections, due to performance reasons.