Implemented `IStateRepository<T>`/`StateRepository<T>` to let the user access the states of the aggregates for debugging or validation purpose.
Implemented the source generation for this, too.

Again: those repos won't be the best choice for production purpose, because you can't filter by properties etc. to get a performant read access to the database! To filter by properties, the complete database has to be loaded into memory (because deserialization is needed to filter).

Only: validating a single aggregate (get by id), maybe for state change, could be useful and somewhat performant.